Disable syscall usage for windows build
