+++ commands/init.go
@@ -10,10 +10,8 @@ import (
 	"net"
 	"net/http"
 	"os"
-	"os/user"
 	"runtime"
 	"strings"
-	"syscall"
 
 	"github.com/mattermost/mattermost-server/v5/model"
 
@@ -159,34 +157,6 @@ func InitWebSocketClient() (*model.WebSocketClient, error) {
 		return nil, errors.Wrap(appErr, "unable to create the websockets connection")
 	}
 	return client, nil
-}
-
-func checkValidSocket(socketPath string) error {
-	// check file mode and permissions
-	fi, err := os.Stat(socketPath)
-	if err != nil && os.IsNotExist(err) {
-		return fmt.Errorf("socket file %q doesn't exists, please check the server configuration for local mode", socketPath)
-	} else if err != nil {
-		return err
-	}
-	if fi.Mode() != expectedSocketMode {
-		return fmt.Errorf("invalid file mode for file %q, it must be a socket with 0600 permissions", socketPath)
-	}
-
-	// check matching user
-	cUser, err := user.Current()
-	if err != nil {
-		return err
-	}
-	s, ok := fi.Sys().(*syscall.Stat_t)
-	if !ok {
-		return fmt.Errorf("cannot get owner of the file %q", socketPath)
-	}
-	if fmt.Sprint(s.Uid) != cUser.Uid {
-		return fmt.Errorf("owner of the file %q must be the same user running mmctl", socketPath)
-	}
-
-	return nil
 }
 
 func InitUnixClient(socketPath string) (*model.Client4, error) {
